+++ app/src/main/java/woowacourse/shopping/presentation/cart/ui/CartViewHolder.kt
@@ -1,9 +1,9 @@
-package woowacourse.shopping.presentation.ui.cart
+package woowacourse.shopping.presentation.cart.ui
 
 import androidx.recyclerview.widget.RecyclerView
 import woowacourse.shopping.databinding.ItemCartBinding
 import woowacourse.shopping.domain.model.CartProduct
-import woowacourse.shopping.presentation.viewmodel.cart.CartViewModel
+import woowacourse.shopping.presentation.cart.CartViewModel
 
 class CartViewHolder(
     private val binding: ItemCartBinding,
