+++ app/src/main/res/layout/activity_products.xml
@@ -10,7 +10,7 @@
             android:id="@+id/cl_products"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
-            tools:context=".presentation.ui.products.ProductsActivity">
+            tools:context=".presentation.products.ui.ProductsActivity">
 
 
             <androidx.recyclerview.widget.RecyclerView
