+++ app/src/main/res/layout/item_last_product.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="vm"
-            type="woowacourse.shopping.presentation.viewmodel.products.ProductsViewModel" />
+            type="woowacourse.shopping.presentation.products.ProductsViewModel" />
 
         <variable
             name="product"
