+++ app/src/test/java/woowacourse/shopping/ui/viewmodel/ProductEntityDetailViewModelTest.kt
@@ -6,7 +6,7 @@ import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.ExtendWith
 import woowacourse.shopping.data.repository.CartDummyRepositoryImpl
 import woowacourse.shopping.data.repository.ProductDummyRepositoryImpl
-import woowacourse.shopping.presentation.viewmodel.productdetail.ProductDetailViewModel
+import woowacourse.shopping.presentation.productdetail.ProductDetailViewModel
 import woowacourse.shopping.util.InstantTaskExecutorExtension
 import woowacourse.shopping.util.getOrAwaitValue
